% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_annotation.R
\name{annotate_sites}
\alias{annotate_sites}
\title{Annotate phosphoproteome with functional features}
\usage{
annotate_sites(phosphoproteome)
}
\arguments{
\item{phosphoproteome}{A data frame with the columns `acc`, `position` and `residue`
referring to a list of phosphorylation sites}
}
\value{
A data frame expanding the input data frame with a list of functional features.
}
\description{
This function takes annotation files contained within the package to annotate a
phosphoproteome with extra functional information.
}
