% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_disopred}
\alias{feature_disopred}
\title{Disorder predictions from disopred}
\format{Data frame with 1.87M rows
\describe{
  \item{acc}{uniprot accession for protein}
  \item{position}{position of the acceptor residue}
  \item{residue}{residue of the acceptor residue}
  \item{is_disopred}{is the residue in a disordered region}
  \item{disopred_score}{score for disopred prediction}
}}
\source{
Ochoa et al. The functional landscape of the human phosphoproteome
}
\usage{
feature_disopred
}
\description{
A dataset containing disorder predictions for every S, T or Y.
}
\keyword{datasets}
