% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_disprot}
\alias{feature_disprot}
\title{Disorder annotations in Disprot}
\format{Data frame with 5,255 rows
\describe{
  \item{acc}{uniprot accession for protein}
  \item{position}{position of the acceptor residue}
  \item{residue}{residue of the acceptor residue}
  \item{disprot_disorder}{is annotated as a disordered region}
}}
\source{
Ochoa et al. The functional landscape of the human phosphoproteome
}
\usage{
feature_disprot
}
\description{
A dataset containing disorder annotations for S, T or Y.
}
\keyword{datasets}
