% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_domains}
\alias{feature_domains}
\title{Disorder annotations in Disprot}
\format{Data frame with 5,255 rows
\describe{
  \item{acc}{uniprot accession for protein}
  \item{position}{position of the acceptor residue}
  \item{residue}{residue of the acceptor residue}
  \item{Coiled_coil_uniprot}{is annotated as a coiled coil in uniprot}
  \item{comp_bias_uniprot}{is annotated as a compositional bias region in uniprot}
  \item{domain_uniprot}{is annotated as a domain in uniprot}
  \item{motif_uniprot}{is annotated as a motif in uniprot}
  \item{region_uniprot}{is annotated as a region in uniprot}
  \item{repeat_uniprot}{is annotated as a repeat region in uniprot}
  \item{zn_finger_uniprot}{is annotated as a Zn finger region in uniprot}
}}
\source{
Ochoa et al. The functional landscape of the human phosphoproteome
}
\usage{
feature_domains
}
\description{
A dataset containing region annotations from Uniprot.
}
\keyword{datasets}
