% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_scaling}
\alias{log_scaling}
\title{Function to log scale resulting functional scores}
\usage{
log_scaling(d)
}
\arguments{
\item{d}{vector of scores}
}
\value{
log-scaled scores
}
\description{
Function to log scale resulting functional scores
}
