% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phosphoproteome}
\alias{phosphoproteome}
\title{Reference human phosphoproteome}
\format{Data frame with 116,258 rows
\describe{
  \item{acc}{uniprot accession for protein}
  \item{position}{position of the acceptor residue}
  \item{residue}{residue of the acceptor residue}
}}
\source{
Ochoa et al. The functional landscape of the human phosphoproteome
}
\usage{
phosphoproteome
}
\description{
List of human phosphorylation sites gathered from a big search of mass spectrometry
data contained in the PRIDE database.
}
\keyword{datasets}
