% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_score.R
\name{predict_funscore}
\alias{predict_funscore}
\title{Predict functional score using trained model}
\usage{
predict_funscore(preprocessed_phos, model, ncores = 1)
}
\arguments{
\item{preprocessed_phos}{preprocessed data frame of features}

\item{model}{trained model}

\item{ncores}{number of cores for parallelized computation. Defauls to 1.}
}
\value{
A data frame of scored phosphosites
}
\description{
Apply model to predict functional score to a data frame of preprocessed features.
}
