% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psp}
\alias{psp}
\title{List of human phosphosites with known regulatory functions}
\format{Data frame with 6,249 rows
\describe{
  \item{acc}{uniprot accession for protein}
  \item{position}{position of the acceptor residue}
}}
\source{
PhosphositePlus \url{https://www.phosphosite.org/}
}
\usage{
psp
}
\description{
This list is the result of downloading and parsing the sites reported in the PhosphositePlus database (see license).
}
\keyword{datasets}
