% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_score.R
\name{train_funscore}
\alias{train_funscore}
\title{Training the phosphosite functional score}
\usage{
train_funscore(preprocessed_phos, residue, gold, parameters = NULL,
  seed = NULL, ncores = 1)
}
\arguments{
\item{preprocessed_phos}{data frame of preprocessed features}

\item{residue}{residue set to train. Options are "ST" or "Y"}

\item{gold}{data frame containing a gold standard set of known functional phosphosites. D
ata frame should contain at least 2 columns: "acc" and "position".}

\item{parameters}{list of parameters with training settings}

\item{seed}{seed to reproduce the training}

\item{ncores}{number of cores to use in parallelized training. Defaults to 1.}
}
\value{
object with the trained model
}
\description{
Function to train a phosphosite functional score based on a phosphoproteome and a provided gold standard.
}
